+++ forge/src/main/java/com/jaquadro/minecraft/storagedrawers/StorageDrawers.java
@@ -76,6 +76,7 @@ public class StorageDrawers
 
     private void setup (final FMLCommonSetupEvent event) {
         CompTierRegistry.INSTANCE.initialize();
+        StorageBlacklist.INSTANCE.initialize();
         PlatformCapabilities.initHandlers();
 
         //oreDictRegistry = new OreDictRegistry();