+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 [12.11.4]
 - Possible fix for restoring detached drawer leaving drawer without a label
 - Prevent drawers with contents being able to stack if contents identical
+- Prevent drawers from being stored in bundles and shulker boxes
+- Added config option to prevent storing filled drawers in drawers
+- Added config option to blacklist items or namespaces from being stored
 
 [12.11.3]
 - NEO/FORGE: Fix controller slave no longer responding to connections